Make it work with PowerShell Core on Linux

+++ CreateAliases.ps1
@@ -1,3 +1,5 @@
+$root = Split-Path -Path $MyInvocation.MyCommand.Definition -Parent
+Import-Module "$root\Modules\posh-alias\Posh-Alias.psd1"
 Set-Alias pester invoke-pester
 Set-Alias psake invoke-psake
 function add {
@@ -29,5 +31,7 @@ if (Get-Command hub -ErrorAction Ignore) {
     Add-Alias git "$($(Get-Command hub).Source)"
 }
 if (Get-Command curl -CommandType Application -ErrorAction Ignore) { #use system curl if available
-    Remove-Item alias:curl
+    if (Get-Alias curl -ErrorAction Ignore) {
+      Remove-Item alias:curl
+    }
 }
